Refactor DevicesTable
Create graphql DevicesTable_DeviceFragment in file that use it's data directly
